Fixed #24219 -- Moved SelectDateWidget together with the other widgets and deprecated django.forms.extras.
Thanks Berker Peksag and Tim Graham for the reviews.
