+++ django/forms/extras/__init__.py
@@ -1,3 +1,12 @@
-from django.forms.extras.widgets import SelectDateWidget
+import warnings
+
+from django.forms.widgets import SelectDateWidget
+from django.utils.deprecation import RemovedInDjango21Warning
 
 __all__ = ['SelectDateWidget']
+
+
+warnings.warn(
+    "django.forms.extras is deprecated. You can find "
+    "SelectDateWidget in django.forms.widgets instead.",
+    RemovedInDjango21Warning, stacklevel=2)
